Delete data for an interval after each power-up event

This trims the spurious data that instruments can produce for a short time interval after each power-on events.

Usage

deleteStartupData(
  x,
  gapTime = 60,
  deleteTime = 20,
  stream = "payload1",
  debug = getOption("gliderDebug", default = 0)
)

Arguments

x

an [oceglider-class] object, e.g. as read by `read.read.glider.seaexplorer.raw()` or similar functions.

gapTime

numeric value, in seconds, used to infer power-on events. There is always one event at the start of the time series. In addition, any time differences that exceed `startupCriterion` seconds are considered to be power-off events, so the first time after such an interval is taken to be a power-on event.

deleteTime

numeric value indicating the time interval, in seconds, of data that will be removed.

stream

character value naming the stream to use, if the object is multi-streamed, meaning that the `data` slot contains multiple data frames, one for the glider (named `glider`) another for the first payload (named `"payload"`), and possibly other payloads. The value of `stream` is ignored unless the object's `metadata` slot contains an item named `dataAreStreamed` and that item is TRUE.

debug

an integer, less than or equal to zero for no message, and positive to print information about the processing.

Author

Dan Kelley