Convert SBE Oxygen Frequency Value to Oxygen Saturation

The calibration formula comes from SBE calibration sheets. This code is based on `sbeO2Hz2Sat()` in https://github.com/DFOglider/pilotingApp, with minor modifications: (1) it has a new name to avoid conflicts, (2) the calibration coefficients are collected into a single parameter, `cal`, (3) `oxygenFrequency` is renamed `frequency`, (3) `unit` is added, to permit specifying the unit of the return value, and (5) the conversion to umol/kg takes into account the density, as indicated in Reference 1.

Usage

swOxygenFrequencyToSaturation(
  temperature,
  salinity,
  pressure,
  frequency,
  cal,
  unit = "ml/l"
)

Arguments

temperature

numeric value holding temperature in degrees C.

salinity

numeric value holding Practical Salinity.

pressure

numeric value holding pressure in dbar.

frequency

the frequency reading from the oxygen sensor system, in Hz.

cal

list or data frame holding calibration parameters named `Soc`, `Foffset`, `A`, `B`, `C`, `D` and `Enom`. The last of these is called "E nominal" in the calibration sheets.

unit

character value indicating the desired unit. The default is `ml/l` but `umol/kg` is also accepted, the latter being computed by multiplying the former by `44.6591*1000/(1000+sigma0)`, where `sigma0` is seawater potential density anomaly in kg/m^3 (see Reference 1, which uses the mole fraction stated in Reference 2).

References

1. Thierry, Virginie, Henry Bittig, Denis Gilbert, Taiyo Kobayashi, Sato Kanako, and Claudia Schmid. "Processing Argo Oxygen Data at the DAC Level (version 2.3.3)". Ifremer, 2022. doi:10.13155/39795

2. Garcia, Herncin E., and Louis I. Gordon. "Oxygen Solubility in Seawater: Better Fitting Equations." Limnology and Oceanography 37, no. 6 (September 1992): 1307–12. doi:10.4319/lo.1992.37.6.1307

Author

Chantelle Layton and Dan Kelley