oceglider: A Package for Processing Ocean Glider Data

This package was written with two particular dataset types in mind, from SeaExplorer and Slocum devices. There is a good chance that the functions provided here (a) will fail on other types and (b) function names and arguments will change as more datasets are examined by the author.

This package was written with two particular dataset types in mind, from SeaExplorer and Slocum devices. There is a good chance that the functions provided here (a) will fail on other types and (b) function names and arguments will change as more datasets are examined by the author.

See also

Author

Maintainer: Dan Kelley dan.kelley@dal.ca (ORCID)

Authors:

Clark Richards clark.richards@gmail.com (ORCID)

Chantelle Layton layton.chantelle@gmail.com (ORCID)