Read a glider file in netcdf format

This is a provisional function, written to handle some particular files available to the author.

read.glider.netcdf(file, debug)

Arguments

file

Name of a netcdf file.

debug

an integer specifying whether debugging information is to be printed during processing. If this is not provided, then the value of getOption("gliderDebug",0) is used. The printing is done by a call to gliderDebug. Setting debug=0 turns off this form of debugging, while higher values may yield more information, depending on the function. If one glider function calls another, it passes the value of debug but decreased by 1, which means that the value of debug controls not just the breadth of debugging, but also the depth.

Value

A glider object, i.e. one inheriting from glider. (This class inherits from oce::oce in the oce package.)

Details

The data are copied directly from the file, except that time is converted from an integer to a POSIX time. Variable names containing underscores are renamed as e.g. profile_direction to profileDirection, although the [[,glider-method mechanism works with either name, e.g. if g is a glider object, then g[["profileDirection"]] and g[["profile_direction"]] give the same result.

See also

Author

Dan Kelley

Examples

if (FALSE) {
library(oceglider)

# NOTE: these files are of order 100Meg, so they are
# not provided with the package as samples. In both
# examples, we plot a map and then an incidence-TS plot.

# Seaexplorer data, from DFO (January 2019)
g <- read.glider.netcdf("~/Dropbox/glider_dfo.nc")
# Remove spurious times, from a year before deployment
g <- subset(g, time > as.POSIXct("2018-01-01"))
# Remove any observation with bad salinity
g <- subset(g, is.finite(g[["salinity"]]))
plot(g, which="map")
ctd <- as.ctd(g[["salinity"]], g[["temperature"]], g[["pressure"]],
              longitude=g[["longitude"]], latitude=g[["latitude"]])
plotTS(ctd, useSmoothScatter=TRUE)

# Slocum data,from Dalhousie CEOTR rdapp (April 2019)
g <- read.glider.netcdf("~/Dropbox/glider_erdapp.nc")
# Remove any observation with bad salinity
g <- subset(g, is.finite(g[["salinity"]]))
plot(g, which="map")
ctd <- as.ctd(g[["salinity"]], g[["temperature"]], g[["pressure"]],
              latitude=g[["latitude"]], longitude=g[["longitude"]])
plotTS(ctd, useSmoothScatter=TRUE)
}